refactor: secret key 디코딩 여러번 -> 한번
토큰을 새로 발급할 때마다 매번 Base64로 인코딩된 secret key를 디코딩해서 사용하는 것에서 생성자에서 한번만 변경해서 해당 값을 이용하는 것으로 변경
